+++ lib/new_relic/harvest/telemetry_sdk/config.ex
@@ -1,0 +1,10 @@
+defmodule NewRelic.Harvest.TelemetrySdk.Config do
+  @moduledoc false
+
+  @default %{
+    logs_harvest_cycle: 5_000
+  }
+  def lookup(key) do
+    Application.get_env(:new_relic_agent, key) || @default[key]
+  end
+end