+++ test/integration/integration_test.exs
@@ -7,12 +7,12 @@ defmodule IntegrationTest do
   # env NR_INT_TEST=true mix test test/integration --include skip
 
   setup do
-    System.put_env("NEW_RELIC_HARVEST_ENABLED", "true")
+    reset_config = TestHelper.update(:nr_config, harvest_enabled: true)
     Collector.AgentRun.reconnect()
     GenServer.call(Collector.AgentRun, :ping)
 
     on_exit(fn ->
-      System.delete_env("NEW_RELIC_HARVEST_ENABLED")
+      reset_config.()
     end)
 
     :ok
@@ -69,7 +69,16 @@ defmodule IntegrationTest do
       Collector.Protocol.metric_data([agent_run_id, ts_start, ts_end, data_array])
   end
 
+  test "Can post a Log" do
+    {:ok, resp} =
+      NewRelic.Harvest.TelemetrySdk.API.log([%{logs: [%{message: "TEST"}], common: %{}}])
+
+    assert resp.status_code == 202
+  end
+
   test "EnabledSupervisor starts" do
+    NewRelic.EnabledSupervisorManager.start_child()
+
     # make sure a process under EnabledSupervisor started
     assert Process.whereis(NewRelic.Sampler.Beam)
   end